WIP benchmark
It feels like I've almost got this, but I've got one last borrow checker
error that I don't understand. It seems like the `capnp_as_ref` borrow
shouldn't be lasting to the end of the `for` loop but it is.

Try a compile with `cargo run -p benchmark` from inside the `capnp`
directory.

The general meta-struct strategy does seem to work in:
https://github.com/capnproto/capnproto-rust/blob/master/benchmark/benchmark.rs